feat: placeholder for account stats
