+++ pages/api/getFollowers.ts
@@ -3,7 +3,8 @@ import { z } from 'zod';
 
 import { prisma } from '@/lib/prismadb';
 import { httpCodes, responseMessages } from '@/utils/apiResponses';
-const USERS_PER_SCROLL = 6;
+
+import { USERS_PER_SCROLL } from '@/components/molecules/statsModal/StatsModal';
 
 const schema = z.object({
   userId: z.string(),
